backend/requirements.txt
# ILMI — Backend Dependencies
# ==========================================

# Django Core
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
django-allauth>=65.0,<66.0
django-argon2-hasher>=1.0,<2.0
argon2-cffi>=23.1,<24.0
pyotp>=2.9,<3.0

# Database
psycopg2-binary>=2.9,<3.0

# Caching & Task Queue
redis>=5.0,<6.0
celery>=5.4,<6.0
django-celery-beat>=2.6,<3.0
django-celery-results>=2.5,<3.0

# Real-time (WebSocket)
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0

# File Storage
django-storages>=1.14,<2.0
boto3>=1.34,<2.0
Pillow>=10.0,<11.0

# QR Code Generation
qrcode>=8.0,<9.0

# PDF Generation
WeasyPrint>=62.0,<63.0

# AI Chatbot (RAG)
langchain>=0.3,<0.4
langchain-openai>=0.2,<0.3
langchain-pinecone>=0.2,<0.3
openai>=1.50,<2.0

# Push Notifications (Firebase V1 API)
firebase-admin>=6.5,<7.0

# Excel Import/Export
openpyxl>=3.1,<4.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Monitoring & Logging
sentry-sdk>=2.0,<3.0

# Utilities
python-decouple>=3.8,<4.0
django-extensions>=3.2,<4.0
whitenoise>=6.7,<7.0

# Development
ipython>=8.0
django-debug-toolbar>=4.4,<5.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0
faker>=30.0,<31.0
